Fixed OrderBy type
