+++ front/src/modules/object-metadata/types/OrderBy.ts
@@ -1,0 +1,5 @@
+export type OrderBy =
+  | 'AscNullsLast'
+  | 'DescNullsLast'
+  | 'AscNullsFirst'
+  | 'DescNullsFirst';
