+++ front/src/modules/object-metadata/utils/getObjectOrderByField.ts
@@ -1,6 +1,6 @@
 import { ObjectMetadataItem } from '@/object-metadata/types/ObjectMetadataItem';
+import { OrderBy } from '@/object-metadata/types/OrderBy';
 import { OrderByField } from '@/object-metadata/types/OrderByField';
-import { OrderBy } from '@/object-record/select/hooks/useRecordsForSelect';
 import { FieldMetadataType } from '~/generated-metadata/graphql';
 
 export const getObjectOrderByField = (
